댓글, 커뮤니티 합침+ 패키지명 소문자로 수정
